requirements.txt
fastapi==0.115.0
hypercorn==0.15.0
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.1.0
python-dotenv==1.0.0
